Refactor codebase into modular structure
- Move core functions into dedicated modules (io.rs, grouping.rs, analysis.rs, types.rs)
- Relocate HanziRecord, HanziOnset, and HanziRime to types.rs
- Update all imports and maintain backward compatibility through lib.rs re-exports
- Preserve all existing functionality and tests
- Improve code organization and maintainability
