+++ src/lib.rs
@@ -23,1302 +23,28 @@
 //! - [`set_hanzi_onsets`]: Analyzes and sets onset information for characters
 //! - [`set_hanzi_rime`]: Analyzes and sets rime information for characters
 
-use std::collections::HashMap;
-use std::io::BufRead;
-use std::str::FromStr;
+pub mod analysis;
+pub mod grouping;
+pub mod io;
+pub mod types;
 
-/// Enumeration of Hanzi onset sounds (initial consonants)
-///
-/// This enum represents all possible onset sounds in Mandarin Chinese pinyin.
-/// Onsets are the initial consonant sounds that begin a syllable. Some syllables
-/// may have no onset (represented by `None`).
-///
-/// # Examples of onsets
-///
-/// - `B`: as in "bā" (八)
-/// - `Zh`: as in "zhōng" (中)
-/// - `None`: as in "ā" (啊) - syllables starting with vowels
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum HanziOnset {
-    B,
-    P,
-    M,
-    F,
-    D,
-    T,
-    N,
-    Z,
-    C,
-    S,
-    L,
-    Zh,
-    Ch,
-    Sh,
-    R,
-    J,
-    Q,
-    X,
-    G,
-    K,
-    H,
-    Y,
-    W,
-    None,
-}
+// Re-export the types module for public API
+pub use crate::types::{HanziOnset, HanziRecord, HanziRime};
 
-impl HanziOnset {
-    /// Returns the kebab-case string representation of the onset
-    ///
-    /// This method converts the onset to a kebab-case string format,
-    /// where compound onsets (like `Zh`, `Ch`, `Sh`) are represented
-    /// in lowercase with no separators, and `None` is represented as "none".
-    ///
-    /// # Returns
-    ///
-    /// A string slice representing the onset in kebab-case format
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use study_rust_hanzi::HanziOnset;
-    ///
-    /// assert_eq!(HanziOnset::B.as_str(), "b");
-    /// assert_eq!(HanziOnset::Zh.as_str(), "zh");
-    /// assert_eq!(HanziOnset::None.as_str(), "none");
-    /// ```
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            HanziOnset::B => "b",
-            HanziOnset::P => "p",
-            HanziOnset::M => "m",
-            HanziOnset::F => "f",
-            HanziOnset::D => "d",
-            HanziOnset::T => "t",
-            HanziOnset::N => "n",
-            HanziOnset::Z => "z",
-            HanziOnset::C => "c",
-            HanziOnset::S => "s",
-            HanziOnset::L => "l",
-            HanziOnset::Zh => "zh",
-            HanziOnset::Ch => "ch",
-            HanziOnset::Sh => "sh",
-            HanziOnset::R => "r",
-            HanziOnset::J => "j",
-            HanziOnset::Q => "q",
-            HanziOnset::X => "x",
-            HanziOnset::G => "g",
-            HanziOnset::K => "k",
-            HanziOnset::H => "h",
-            HanziOnset::Y => "y",
-            HanziOnset::W => "w",
-            HanziOnset::None => "none",
-        }
-    }
-}
+// Re-export the io module functions for backward compatibility
+pub use crate::io::read_hanzi_file;
 
-impl std::str::FromStr for HanziOnset {
-    type Err = String;
+// Re-export the grouping module functions for backward compatibility
+pub use crate::grouping::{
+    format_pinyin_output, format_tone_output, group_by_pinyin, group_by_tone,
+};
 
-    /// Parses a string into a HanziOnset
-    ///
-    /// This method converts a string representation back into a HanziOnset variant.
-    /// It accepts both the exact case-sensitive variants and their lowercase equivalents.
-    ///
-    /// # Arguments
-    ///
-    /// * `s` - The string to parse
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(HanziOnset)` - If the string matches a valid onset
-    /// * `Err(String)` - If the string doesn't match any known onset
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::str::FromStr;
-    /// use study_rust_hanzi::HanziOnset;
-    ///
-    /// assert_eq!(HanziOnset::from_str("b"), Ok(HanziOnset::B));
-    /// assert_eq!(HanziOnset::from_str("zh"), Ok(HanziOnset::Zh));
-    /// assert_eq!(HanziOnset::from_str("none"), Ok(HanziOnset::None));
-    /// assert!(HanziOnset::from_str("invalid").is_err());
-    /// ```
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "b" => Ok(HanziOnset::B),
-            "p" => Ok(HanziOnset::P),
-            "m" => Ok(HanziOnset::M),
-            "f" => Ok(HanziOnset::F),
-            "d" => Ok(HanziOnset::D),
-            "t" => Ok(HanziOnset::T),
-            "n" => Ok(HanziOnset::N),
-            "z" => Ok(HanziOnset::Z),
-            "c" => Ok(HanziOnset::C),
-            "s" => Ok(HanziOnset::S),
-            "l" => Ok(HanziOnset::L),
-            "zh" => Ok(HanziOnset::Zh),
-            "ch" => Ok(HanziOnset::Ch),
-            "sh" => Ok(HanziOnset::Sh),
-            "r" => Ok(HanziOnset::R),
-            "j" => Ok(HanziOnset::J),
-            "q" => Ok(HanziOnset::Q),
-            "x" => Ok(HanziOnset::X),
-            "g" => Ok(HanziOnset::G),
-            "k" => Ok(HanziOnset::K),
-            "h" => Ok(HanziOnset::H),
-            "y" => Ok(HanziOnset::Y),
-            "w" => Ok(HanziOnset::W),
-            "none" => Ok(HanziOnset::None),
-            _ => Err(format!("Invalid onset: '{s}'")),
-        }
-    }
-}
-
-/// Enumeration of Hanzi rime sounds (vowels and final consonants)
-///
-/// This enum represents all possible rime sounds in Mandarin Chinese pinyin.
-/// Rimes are the vowel and optional final consonant parts of a syllable that
-/// follow the onset. Every syllable must have a rime.
-///
-/// # Examples of rimes
-///
-/// - `A`: as in "mā" (妈) - simple vowel
-/// - `Ang`: as in "tāng" (汤) - vowel + nasal consonant
-/// - `Iang`: as in "liáng" (良) - complex vowel + nasal
-/// - `V`: represents "ü" as in "nǚ" (女)
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum HanziRime {
-    E,
-    A,
-    Ei,
-    Ai,
-    Ou,
-    Ao,
-    En,
-    An,
-    Ong,
-    Eng,
-    Ang,
-    I,
-    Ie,
-    Ia,
-    Iu,
-    Iao,
-    In,
-    Ian,
-    Iong,
-    Ing,
-    Iang,
-    U,
-    Uo,
-    Ua,
-    Ui,
-    Uai,
-    Un,
-    Uan,
-    Uang,
-    V,
-    Ve,
-    None,
-}
-
-impl HanziRime {
-    /// Returns the string representation of the rime
-    ///
-    /// This method converts the rime to its pinyin string representation.
-    /// The special case `V` is converted to "ü" and `Ve` to "üe" to represent
-    /// the umlaut vowel sounds in Mandarin pinyin.
-    ///
-    /// # Returns
-    ///
-    /// A string slice representing the rime in pinyin format
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use study_rust_hanzi::HanziRime;
-    ///
-    /// assert_eq!(HanziRime::A.as_str(), "a");
-    /// assert_eq!(HanziRime::Ang.as_str(), "ang");
-    /// assert_eq!(HanziRime::V.as_str(), "ü");
-    /// assert_eq!(HanziRime::Ve.as_str(), "üe");
-    /// assert_eq!(HanziRime::None.as_str(), "none");
-    /// ```
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            HanziRime::E => "e",
-            HanziRime::A => "a",
-            HanziRime::Ei => "ei",
-            HanziRime::Ai => "ai",
-            HanziRime::Ou => "ou",
-            HanziRime::Ao => "ao",
-            HanziRime::En => "en",
-            HanziRime::An => "an",
-            HanziRime::Ong => "ong",
-            HanziRime::Eng => "eng",
-            HanziRime::Ang => "ang",
-            HanziRime::I => "i",
-            HanziRime::Ie => "ie",
-            HanziRime::Ia => "ia",
-            HanziRime::Iu => "iu",
-            HanziRime::Iao => "iao",
-            HanziRime::In => "in",
-            HanziRime::Ian => "ian",
-            HanziRime::Iong => "iong",
-            HanziRime::Ing => "ing",
-            HanziRime::Iang => "iang",
-            HanziRime::U => "u",
-            HanziRime::Uo => "uo",
-            HanziRime::Ua => "ua",
-            HanziRime::Ui => "ui",
-            HanziRime::Uai => "uai",
-            HanziRime::Un => "un",
-            HanziRime::Uan => "uan",
-            HanziRime::Uang => "uang",
-            HanziRime::V => "ü",
-            HanziRime::Ve => "üe",
-            HanziRime::None => "none",
-        }
-    }
-}
-
-impl std::str::FromStr for HanziRime {
-    type Err = String;
-
-    /// Parses a string into a HanziRime
-    ///
-    /// This method converts a string representation back into a HanziRime variant.
-    /// It accepts both the exact pinyin representation and handles special cases
-    /// like "ü" and "üe" for the V and Ve variants.
-    ///
-    /// # Arguments
-    ///
-    /// * `s` - The string to parse
-    ///
-    /// # Returns
-    ///
-    /// * `Ok(HanziRime)` - If the string matches a valid rime
-    /// * `Err(String)` - If the string doesn't match any known rime
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use std::str::FromStr;
-    /// use study_rust_hanzi::HanziRime;
-    ///
-    /// assert_eq!(HanziRime::from_str("a"), Ok(HanziRime::A));
-    /// assert_eq!(HanziRime::from_str("ang"), Ok(HanziRime::Ang));
-    /// assert_eq!(HanziRime::from_str("ü"), Ok(HanziRime::V));
-    /// assert_eq!(HanziRime::from_str("üe"), Ok(HanziRime::Ve));
-    /// assert_eq!(HanziRime::from_str("none"), Ok(HanziRime::None));
-    /// assert!(HanziRime::from_str("invalid").is_err());
-    /// ```
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "e" => Ok(HanziRime::E),
-            "a" => Ok(HanziRime::A),
-            "ei" => Ok(HanziRime::Ei),
-            "ai" => Ok(HanziRime::Ai),
-            "ou" => Ok(HanziRime::Ou),
-            "ao" => Ok(HanziRime::Ao),
-            "en" => Ok(HanziRime::En),
-            "an" => Ok(HanziRime::An),
-            "ong" => Ok(HanziRime::Ong),
-            "eng" => Ok(HanziRime::Eng),
-            "ang" => Ok(HanziRime::Ang),
-            "i" => Ok(HanziRime::I),
-            "ie" => Ok(HanziRime::Ie),
-            "ia" => Ok(HanziRime::Ia),
-            "iu" => Ok(HanziRime::Iu),
-            "iao" => Ok(HanziRime::Iao),
-            "in" => Ok(HanziRime::In),
-            "ian" => Ok(HanziRime::Ian),
-            "iong" => Ok(HanziRime::Iong),
-            "ing" => Ok(HanziRime::Ing),
-            "iang" => Ok(HanziRime::Iang),
-            "u" => Ok(HanziRime::U),
-            "uo" => Ok(HanziRime::Uo),
-            "ua" => Ok(HanziRime::Ua),
-            "ui" => Ok(HanziRime::Ui),
-            "uai" => Ok(HanziRime::Uai),
-            "un" => Ok(HanziRime::Un),
-            "uan" => Ok(HanziRime::Uan),
-            "uang" => Ok(HanziRime::Uang),
-            "ü" => Ok(HanziRime::V),
-            "üe" => Ok(HanziRime::Ve),
-            "v" => Ok(HanziRime::V),   // Alternative representation for ü
-            "ve" => Ok(HanziRime::Ve), // Alternative representation for üe
-            "none" => Ok(HanziRime::None),
-            _ => Err(format!("Invalid rime: '{s}'")),
-        }
-    }
-}
-
-/// Represents a single Chinese character with all its linguistic and frequency data
-///
-/// This structure contains comprehensive information about a Chinese character,
-/// including both simplified and traditional forms, pinyin pronunciation data,
-/// frequency information, and phonetic analysis (onset/rime breakdown).
-///
-/// # Fields
-///
-/// * `frequency` - Frequency rank of the character (lower numbers = more common)
-/// * `simplified` - Simplified Chinese character form
-/// * `traditional` - Traditional Chinese character form  
-/// * `pinyin` - Complete pinyin with tone marks (e.g., "mā")
-/// * `pinyin_without_tone` - Pinyin without tone marks (e.g., "ma")
-/// * `tone` - Tone number (1-4 for tones, 5 for neutral tone)
-/// * `onset` - Initial consonant sound classification
-/// * `rime` - Vowel and final consonant sound classification
-pub struct HanziRecord {
-    pub frequency: u32,
-    pub simplified: String,
-    pub traditional: String,
-    pub pinyin: String,
-    pub pinyin_without_tone: String,
-    pub tone: u32,
-    pub onset: HanziOnset,
-    pub rime: HanziRime,
-}
-
-/// Reads a TSV file containing Hanzi data and returns a vector of HanziRecord
-///
-/// This function parses a tab-separated values file where each line represents
-/// one Chinese character with its associated data. The expected format is:
-/// frequency, simplified, traditional, pinyin, pinyin_without_tone, tone
-///
-/// # Arguments
-///
-/// * `file_path` - Path to the TSV file to read
-///
-/// # Returns
-///
-/// * `Ok(Vec<HanziRecord>)` - Successfully parsed records
-/// * `Err(std::io::Error)` - File I/O error occurred
-///
-/// # File Format
-///
-/// Each line should contain 6 tab-separated fields:
-/// 1. Frequency rank (integer)
-/// 2. Simplified character (string)
-/// 3. Traditional character (string)  
-/// 4. Pinyin with tone marks (string)
-/// 5. Pinyin without tone marks (string)
-/// 6. Tone number (integer, 1-5)
-///
-/// Lines with fewer than 6 fields are skipped. Invalid numbers default to 0.
-pub fn read_hanzi_file(file_path: &str) -> std::io::Result<Vec<HanziRecord>> {
-    let mut records = Vec::new();
-    let file = std::fs::File::open(file_path)?;
-    let reader = std::io::BufReader::new(file);
-
-    for line in reader.lines() {
-        let line = line?;
-        let parts: Vec<&str> = line.split('\t').collect();
-        if parts.len() < 6 {
-            continue; // Skip lines that do not have enough fields
-        }
-        let record = HanziRecord {
-            frequency: parts[0].parse().unwrap_or(0),
-            simplified: parts[1].to_string(),
-            traditional: parts[2].to_string(),
-            pinyin: parts[3].to_string(),
-            pinyin_without_tone: parts[4].to_string(),
-            tone: parts[5].parse().unwrap_or(0),
-            onset: HanziOnset::None, // Set as initial value
-            rime: HanziRime::None,   // Set as initial value
-        };
-        records.push(record);
-    }
-    Ok(records)
-}
-
-/// Analyzes and sets the onset (initial consonant) for each character's pinyin
-///
-/// This function examines the `pinyin_without_tone` field of each record and
-/// determines the appropriate onset classification based on the initial consonant(s).
-/// The onset field is updated in-place for each record.
-///
-/// The function uses `HanziOnset::from_str()` internally for efficient onset detection,
-/// checking possible onsets in order of decreasing length to ensure proper matching
-/// (e.g., "zh" before "z").
-///
-/// # Arguments
-///
-/// * `records` - Mutable slice of HanziRecord to analyze
-///
-/// # Onset Detection Rules
-///
-/// - Multi-character onsets (zh, ch, sh) are checked first
-/// - Single-character onsets are checked next
-/// - If no onset matches, `HanziOnset::None` is assigned (vowel-initial syllables)
-///
-/// # Examples
-///
-/// - "zhong" → `HanziOnset::Zh`
-/// - "ma" → `HanziOnset::M`  
-/// - "an" → `HanziOnset::None`
-pub fn set_hanzi_onsets(records: &mut [HanziRecord]) {
-    use std::str::FromStr;
-
-    // Define onset candidates in order of decreasing length to ensure proper matching
-    // (e.g., "zh" must be checked before "z")
-    const ONSET_CANDIDATES: &[&str] = &[
-        "zh", "ch", "sh", // Multi-character onsets first
-        "b", "p", "m", "f", "d", "t", "n", "z", "c", "s", "l", "r", "j", "q", "x", "g", "k", "h",
-        "y", "w",
-    ];
-
-    for record in records.iter_mut() {
-        let pinyin = &record.pinyin_without_tone;
-
-        // Try to find the first matching onset
-        record.onset = ONSET_CANDIDATES
-            .iter()
-            .find(|&&onset_str| pinyin.starts_with(onset_str))
-            .and_then(|&onset_str| HanziOnset::from_str(onset_str).ok())
-            .unwrap_or(HanziOnset::None);
-    }
-}
-
-/// Analyzes and sets the rime (vowel + final consonant) for each character's pinyin
-///
-/// This function determines the rime part of each character's pronunciation by
-/// removing the onset and matching the remaining sound to known rime patterns.
-/// The rime field is updated in-place for each record.
-///
-/// # Arguments
-///
-/// * `records` - Mutable slice of HanziRecord to analyze
-///
-/// # Prerequisites
-///
-/// This function should be called after `set_hanzi_onsets()` since it relies on
-/// the onset field being correctly set to determine the rime portion.
-///
-/// # Rime Detection Process
-///
-/// 1. Gets the string representation of the onset
-/// 2. Strips the onset from the pinyin to isolate the rime part
-/// 3. Matches the rime part against known rime patterns
-/// 4. Sets `HanziRime::None` if no pattern matches
-///
-/// # Examples
-///
-/// - "ma" (onset: M) → rime part "a" → `HanziRime::A`
-/// - "zhong" (onset: Zh) → rime part "ong" → `HanziRime::Ong`
-/// - "nü" (onset: N) → rime part "ü" → `HanziRime::V`
-pub fn set_hanzi_rime(records: &mut [HanziRecord]) {
-    for record in records.iter_mut() {
-        let pinyin = &record.pinyin_without_tone;
-
-        // Get onset string representation
-        let onset_str = record.onset.as_str();
-
-        // Extract rime part excluding onset
-        let rime_part = if onset_str == "none" {
-            pinyin.as_str()
-        } else if let Some(stripped) = pinyin.strip_prefix(onset_str) {
-            stripped
-        } else {
-            // If onset doesn't match, treat the whole string as rime part
-            pinyin.as_str()
-        };
-
-        // Try to parse rime part using HanziRime::from_str()
-        record.rime = HanziRime::from_str(rime_part).unwrap_or(HanziRime::None);
-    }
-}
-
-/// Groups Hanzi records by pinyin without tone marks
-///
-/// Takes a slice of HanziRecord and groups them by their pinyin_without_tone field.
-/// Returns a vector of tuples containing the pinyin and a vector of characters.
-/// The results are sorted by frequency (descending) and then by pinyin (ascending).
-///
-/// # Arguments
-///
-/// * `records` - A slice of HanziRecord to group
-/// * `use_traditional` - Whether to use traditional characters instead of simplified
-///
-/// # Returns
-///
-/// A vector of tuples where each tuple contains:
-/// - The pinyin without tone as a String
-/// - A vector of character strings corresponding to that pinyin
-///
-/// # Sorting Order
-///
-/// Results are sorted by:
-/// 1. Number of characters (descending) - most common pinyin first
-/// 2. Pinyin alphabetically (ascending) - consistent ordering for same frequency
-///
-/// # Examples
-///
-/// ```rust
-/// # use study_rust_hanzi::{HanziRecord, HanziOnset, HanziRime, group_by_pinyin};
-/// # let records = vec![]; // Placeholder for actual records
-/// let grouped = group_by_pinyin(&records, false); // Use simplified characters
-/// // Result: [("de", vec!["的", "得", "地"]), ("yi", vec!["一", "以"]), ...]
-/// ```
-pub fn group_by_pinyin(
-    records: &[HanziRecord],
-    use_traditional: bool,
-) -> Vec<(String, Vec<String>)> {
-    use std::collections::HashMap;
-
-    let mut pinyin_groups: HashMap<&str, Vec<&str>> = HashMap::new();
-    for record in records {
-        let character = if use_traditional {
-            &record.traditional
-        } else {
-            &record.simplified
-        };
-        pinyin_groups
-            .entry(&record.pinyin_without_tone)
-            .or_default()
-            .push(character);
-    }
-
-    // Sort by frequency (descending) and then by pinyin (ascending)
-    let mut sorted_pinyins: Vec<_> = pinyin_groups.iter().collect();
-    sorted_pinyins.sort_by(|a, b| b.1.len().cmp(&a.1.len()).then(a.0.cmp(b.0)));
-
-    sorted_pinyins
-        .into_iter()
-        .map(|(pinyin, characters)| {
-            (
-                pinyin.to_string(),
-                characters.iter().map(|s| s.to_string()).collect(),
-            )
-        })
-        .collect()
-}
-
-/// Formats pinyin grouping data for display with optional line folding
-///
-/// Takes grouped pinyin data and formats it for display, with optional line folding
-/// for long character lists. Each line shows the pinyin, character count, and characters.
-///
-/// # Arguments
-///
-/// * `grouped_data` - A slice of tuples containing pinyin and character vectors
-/// * `fold_size` - Optional width for line folding. If provided, long character lists
-///   will be folded to this width with continuation lines
-///
-/// # Returns
-///
-/// A vector of formatted strings ready for display
-///
-/// # Output Format
-///
-/// Without folding:
-/// ```text
-/// pinyin  :  42 characters_here
-/// ```
-///
-/// With folding (fold_size = 10):
-/// ```text
-/// pinyin  :  42 first_10_ch
-///               next_chars
-/// ```
-///
-/// # Formatting Details
-///
-/// - Pinyin is left-aligned in an 8-character field
-/// - Character count is right-aligned in a 3-character field
-/// - Continuation lines are indented with 14 spaces to align with characters
-pub fn format_pinyin_output(
-    grouped_data: &[(String, Vec<String>)],
-    fold_size: Option<usize>,
-) -> Vec<String> {
-    let mut output_lines = Vec::new();
-
-    for (pinyin, characters) in grouped_data {
-        let char_list = characters.join("");
-
-        if let Some(fold_size) = fold_size {
-            if char_list.len() > fold_size {
-                // Fold long lines: first fold_size chars on the same line as count
-                let chars: Vec<char> = char_list.chars().collect();
-                let first_chunk: String = chars.iter().take(fold_size).collect();
-
-                output_lines.push(format!(
-                    "{:<8}: {:3} {}",
-                    pinyin,
-                    characters.len(),
-                    first_chunk
-                ));
-
-                // Remaining characters in chunks of fold_size
-                for chunk in chars
-                    .iter()
-                    .skip(fold_size)
-                    .collect::<Vec<_>>()
-                    .chunks(fold_size)
-                {
-                    let chunk_str: String = chunk.iter().map(|c| **c).collect();
-                    output_lines.push(format!("              {chunk_str}"));
-                }
-            } else {
-                output_lines.push(format!(
-                    "{:<8}: {:3} {}",
-                    pinyin,
-                    characters.len(),
-                    char_list
-                ));
-            }
-        } else {
-            output_lines.push(format!(
-                "{:<8}: {:3} {}",
-                pinyin,
-                characters.len(),
-                char_list
-            ));
-        }
-    }
-
-    output_lines
-}
-
-/// Groups Hanzi records by tone for a specific pinyin
-///
-/// Filters records by the target pinyin and groups them by tone.
-/// Returns None if no matching records are found.
-///
-/// # Arguments
-///
-/// * `records` - A slice of HanziRecord to search through
-/// * `target_pinyin` - The pinyin (without tone) to filter by
-/// * `use_traditional` - Whether to use traditional characters instead of simplified
-///
-/// # Returns
-///
-/// An optional vector of tuples where each tuple contains:
-/// - The tone number (u32): 1-4 for standard tones, 5 for neutral tone
-/// - The pinyin with tone marks as a String
-/// - A vector of character strings for that tone
-///
-/// Returns `None` if no characters match the target pinyin.
-///
-/// # Tone Sorting
-///
-/// Results are sorted by tone number (1, 2, 3, 4, 5) in ascending order.
-///
-/// # Examples
-///
-/// ```rust
-/// # use study_rust_hanzi::{HanziRecord, HanziOnset, HanziRime, group_by_tone};
-/// # let records = vec![]; // Placeholder for actual records
-/// if let Some(tone_groups) = group_by_tone(&records, "ma", false) {
-///     // tone_groups: [(1, "mā", vec!["妈"]), (3, "mǎ", vec!["马"]), ...]
-/// }
-/// ```
-pub fn group_by_tone(
-    records: &[HanziRecord],
-    target_pinyin: &str,
-    use_traditional: bool,
-) -> Option<Vec<(u32, String, Vec<String>)>> {
-    let matching_records: Vec<_> = records
-        .iter()
-        .filter(|record| record.pinyin_without_tone == target_pinyin)
-        .collect();
-
-    if matching_records.is_empty() {
-        return None;
-    }
-
-    let mut tone_groups: HashMap<u32, (Vec<&str>, &str)> = HashMap::new();
-    for record in matching_records {
-        let character = if use_traditional {
-            &record.traditional
-        } else {
-            &record.simplified
-        };
-        let entry = tone_groups
-            .entry(record.tone)
-            .or_insert_with(|| (Vec::new(), &record.pinyin));
-        entry.0.push(character);
-    }
-
-    // Sort by tone (1, 2, 3, 4, 5 for neutral tone)
-    let mut sorted_tones: Vec<_> = tone_groups.iter().collect();
-    sorted_tones.sort_by_key(|&(tone, _)| *tone);
-
-    Some(
-        sorted_tones
-            .into_iter()
-            .map(|(tone, (characters, pinyin))| {
-                (
-                    *tone,
-                    pinyin.to_string(),
-                    characters.iter().map(|s| s.to_string()).collect(),
-                )
-            })
-            .collect(),
-    )
-}
-
-/// Formats tone grouping data for display
-///
-/// Takes grouped tone data and formats it for display. Each line shows the pinyin
-/// with tone marks followed by the corresponding characters for that tone.
-///
-/// # Arguments
-///
-/// * `tone_groups` - A slice of tuples containing tone data where each tuple has:
-///   - Tone number (u32): 1-4 for standard tones, 5 for neutral tone
-///   - Pinyin with tone marks (String): e.g., "jī", "jí", "jǐ", "jì"
-///   - Character vector (`Vec<String>`): characters with that pinyin and tone
-///
-/// # Returns
-///
-/// A vector of formatted strings ready for display, one per tone group
-///
-/// # Output Format
-///
-/// Each line follows the pattern:
-/// ```text
-/// pinyin_with_tone: characters
-/// ```
-///
-/// # Examples
-///
-/// ```rust
-/// # use study_rust_hanzi::format_tone_output;
-/// let tone_data = vec![
-///     (1, "mā".to_string(), vec!["妈".to_string()]),
-///     (3, "mǎ".to_string(), vec!["马".to_string(), "码".to_string()]),
-/// ];
-/// let output = format_tone_output(&tone_data);
-/// // Result: ["mā: 妈", "mǎ: 马码"]
-/// ```
-///
-/// # Usage with group_by_tone
-///
-/// This function is typically used in conjunction with [`group_by_tone`]:
-/// ```rust,no_run
-/// # use study_rust_hanzi::{group_by_tone, format_tone_output};
-/// # let records = vec![]; // Placeholder
-/// if let Some(tone_groups) = group_by_tone(&records, "ma", false) {
-///     let formatted = format_tone_output(&tone_groups);
-///     for line in formatted {
-///         println!("{}", line);
-///     }
-/// }
-/// ```
-pub fn format_tone_output(tone_groups: &[(u32, String, Vec<String>)]) -> Vec<String> {
-    tone_groups
-        .iter()
-        .map(|(_tone, pinyin, characters)| {
-            let char_list = characters.join("");
-            format!("{pinyin}: {char_list}")
-        })
-        .collect()
-}
+// Re-export the analysis module functions for backward compatibility
+pub use crate::analysis::{set_hanzi_onsets, set_hanzi_rime};
 
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_read_hanzi_file_length() {
-        let result = read_hanzi_file("hanzi.tsv");
-        assert!(result.is_ok(), "Failed to read hanzi.tsv file");
-
-        let records = result.unwrap();
-        assert_eq!(
-            records.len(),
-            5000,
-            "Expected 5000 records in hanzi.tsv, got {}",
-            records.len()
-        );
-    }
-
-    #[test]
-    fn test_read_hanzi_file_tenth_element() {
-        let result = read_hanzi_file("hanzi.tsv");
-        assert!(result.is_ok(), "Failed to read hanzi.tsv file");
-
-        let records = result.unwrap();
-        assert!(
-            records.len() >= 10,
-            "Not enough records in hanzi.tsv, need at least 10"
-        );
-
-        let tenth_record = &records[9]; // 10th element (index 9)
-        assert_eq!(
-            tenth_record.frequency, 10,
-            "Expected frequency 10, got {}",
-            tenth_record.frequency
-        );
-        assert_eq!(
-            tenth_record.simplified, "他",
-            "Expected simplified '他', got '{}'",
-            tenth_record.simplified
-        );
-        assert_eq!(
-            tenth_record.traditional, "他",
-            "Expected traditional '他', got '{}'",
-            tenth_record.traditional
-        );
-        assert_eq!(
-            tenth_record.pinyin, "tā",
-            "Expected pinyin 'tā', got '{}'",
-            tenth_record.pinyin
-        );
-        assert_eq!(
-            tenth_record.pinyin_without_tone, "ta",
-            "Expected pinyin_without_tone 'ta', got '{}'",
-            tenth_record.pinyin_without_tone
-        );
-        assert_eq!(
-            tenth_record.tone, 1,
-            "Expected tone 1, got {}",
-            tenth_record.tone
-        );
-    }
-
-    #[test]
-    fn test_read_hanzi_file_last_element() {
-        let result = read_hanzi_file("hanzi.tsv");
-        assert!(result.is_ok(), "Failed to read hanzi.tsv file");
-
-        let records = result.unwrap();
-        assert_eq!(records.len(), 5000, "Expected exactly 5000 records");
-
-        let last_record = &records[4999]; // 5000th element (index 4999)
-        assert_eq!(
-            last_record.frequency, 5000,
-            "Expected frequency 5000, got {}",
-            last_record.frequency
-        );
-        assert_eq!(
-            last_record.simplified, "鸨",
-            "Expected simplified '鸨', got '{}'",
-            last_record.simplified
-        );
-        assert_eq!(
-            last_record.traditional, "鴇",
-            "Expected traditional '鴇', got '{}'",
-            last_record.traditional
-        );
-        assert_eq!(
-            last_record.pinyin, "bǎo",
-            "Expected pinyin 'bǎo', got '{}'",
-            last_record.pinyin
-        );
-        assert_eq!(
-            last_record.pinyin_without_tone, "bao",
-            "Expected pinyin_without_tone 'bao', got '{}'",
-            last_record.pinyin_without_tone
-        );
-        assert_eq!(
-            last_record.tone, 3,
-            "Expected tone 3, got {}",
-            last_record.tone
-        );
-    }
-
-    #[test]
-    fn test_set_hanzi_onsets() {
-        let result = read_hanzi_file("hanzi.tsv");
-        assert!(result.is_ok(), "Failed to read hanzi.tsv file");
-
-        let mut records = result.unwrap();
-        set_hanzi_onsets(&mut records);
-
-        // All HanziOnset values other than none should appear
-        use std::collections::HashSet;
-        let mut found_onsets = HashSet::new();
-
-        for record in &records {
-            found_onsets.insert(&record.onset);
-        }
-
-        // Define all HanziOnset values except none
-        let expected_onsets = vec![
-            HanziOnset::B,
-            HanziOnset::P,
-            HanziOnset::M,
-            HanziOnset::F,
-            HanziOnset::D,
-            HanziOnset::T,
-            HanziOnset::N,
-            HanziOnset::Z,
-            HanziOnset::C,
-            HanziOnset::S,
-            HanziOnset::L,
-            HanziOnset::Zh,
-            HanziOnset::Ch,
-            HanziOnset::Sh,
-            HanziOnset::R,
-            HanziOnset::J,
-            HanziOnset::Q,
-            HanziOnset::X,
-            HanziOnset::G,
-            HanziOnset::K,
-            HanziOnset::H,
-            HanziOnset::Y,
-            HanziOnset::W,
-        ];
-
-        for expected_onset in &expected_onsets {
-            assert!(
-                found_onsets.contains(expected_onset),
-                "HanziOnset::{expected_onset:?} was not found in any record"
-            );
-        }
-    }
-
-    #[test]
-    fn test_set_hanzi_rime() {
-        let result = read_hanzi_file("hanzi.tsv");
-        assert!(result.is_ok(), "Failed to read hanzi.tsv file");
-
-        let mut records = result.unwrap();
-
-        // First set onset, then set rime
-        set_hanzi_onsets(&mut records);
-        set_hanzi_rime(&mut records);
-
-        // All HanziRime values other than none should appear
-        use std::collections::HashSet;
-        let mut found_rimes = HashSet::new();
-
-        for record in &records {
-            found_rimes.insert(&record.rime);
-        }
-
-        // Define all HanziRime values except none
-        let expected_rimes = vec![
-            HanziRime::E,
-            HanziRime::A,
-            HanziRime::Ei,
-            HanziRime::Ai,
-            HanziRime::Ou,
-            HanziRime::Ao,
-            HanziRime::En,
-            HanziRime::An,
-            HanziRime::Ong,
-            HanziRime::Eng,
-            HanziRime::Ang,
-            HanziRime::I,
-            HanziRime::Ie,
-            HanziRime::Ia,
-            HanziRime::Iu,
-            HanziRime::Iao,
-            HanziRime::In,
-            HanziRime::Ian,
-            HanziRime::Iong,
-            HanziRime::Ing,
-            HanziRime::Iang,
-            HanziRime::U,
-            HanziRime::Uo,
-            HanziRime::Ua,
-            HanziRime::Ui,
-            HanziRime::Uai,
-            HanziRime::Un,
-            HanziRime::Uan,
-            HanziRime::Uang,
-            HanziRime::V,
-            HanziRime::Ve,
-        ];
-
-        // To identify rimes that are not found
-        let mut missing_rimes = Vec::new();
-        for expected_rime in &expected_rimes {
-            if !found_rimes.contains(expected_rime) {
-                missing_rimes.push(expected_rime);
-            }
-        }
-
-        if !missing_rimes.is_empty() {
-            println!("Missing rimes: {missing_rimes:?}");
-            println!(
-                "Found {} unique rimes out of {} expected",
-                found_rimes.len(),
-                expected_rimes.len()
-            );
-
-            // Display the rimes that were actually found
-            let mut found_list: Vec<_> = found_rimes.iter().collect();
-            found_list.sort_by_key(|r| format!("{r:?}"));
-            println!("Found rimes: {found_list:?}");
-        }
-
-        // If there are rimes not found, skip the test or adjust expectations
-        // For now, only check rimes that actually exist
-        for expected_rime in &expected_rimes {
-            if found_rimes.contains(expected_rime) {
-                // Assert success only if it exists
-                continue;
-            } else {
-                // Only warning if it doesn't exist
-                println!("HanziRime::{expected_rime:?} was not found in any record");
-            }
-        }
-    }
-
-    fn create_test_records() -> Vec<HanziRecord> {
-        vec![
-            HanziRecord {
-                frequency: 1,
-                simplified: "机".to_string(),
-                traditional: "機".to_string(),
-                pinyin: "jī".to_string(),
-                pinyin_without_tone: "ji".to_string(),
-                tone: 1,
-                onset: HanziOnset::J,
-                rime: HanziRime::I,
-            },
-            HanziRecord {
-                frequency: 2,
-                simplified: "计".to_string(),
-                traditional: "計".to_string(),
-                pinyin: "jì".to_string(),
-                pinyin_without_tone: "ji".to_string(),
-                tone: 4,
-                onset: HanziOnset::J,
-                rime: HanziRime::I,
-            },
-            HanziRecord {
-                frequency: 3,
-                simplified: "马".to_string(),
-                traditional: "馬".to_string(),
-                pinyin: "mǎ".to_string(),
-                pinyin_without_tone: "ma".to_string(),
-                tone: 3,
-                onset: HanziOnset::M,
-                rime: HanziRime::A,
-            },
-        ]
-    }
-
-    #[test]
-    fn test_group_by_pinyin_simplified() {
-        let records = create_test_records();
-        let grouped = group_by_pinyin(&records, false);
-
-        // ji should come first as it has more characters than ma
-        assert_eq!(grouped[0].0, "ji");
-        assert_eq!(grouped[0].1, vec!["机", "计"]);
-        assert_eq!(grouped[1].0, "ma");
-        assert_eq!(grouped[1].1, vec!["马"]);
-    }
-
-    #[test]
-    fn test_group_by_pinyin_traditional() {
-        let records = create_test_records();
-        let grouped = group_by_pinyin(&records, true);
-
-        // Traditional characters should be used
-        assert_eq!(grouped[0].0, "ji");
-        assert_eq!(grouped[0].1, vec!["機", "計"]);
-        assert_eq!(grouped[1].0, "ma");
-        assert_eq!(grouped[1].1, vec!["馬"]);
-    }
-
-    #[test]
-    fn test_format_pinyin_output_no_fold() {
-        let test_data = vec![
-            ("ji".to_string(), vec!["机".to_string(), "计".to_string()]),
-            ("ma".to_string(), vec!["马".to_string()]),
-        ];
-
-        let output = format_pinyin_output(&test_data, None);
-
-        assert_eq!(output.len(), 2);
-        assert!(output[0].contains("ji"));
-        assert!(output[0].contains("2"));
-        assert!(output[0].contains("机计"));
-        assert!(output[1].contains("ma"));
-        assert!(output[1].contains("1"));
-        assert!(output[1].contains("马"));
-    }
-
-    #[test]
-    fn test_format_pinyin_output_with_fold() {
-        let test_data = vec![(
-            "test".to_string(),
-            vec![
-                "一".to_string(),
-                "二".to_string(),
-                "三".to_string(),
-                "四".to_string(),
-                "五".to_string(),
-            ],
-        )];
-
-        let output = format_pinyin_output(&test_data, Some(3));
-
-        // fold_size is 3, so first line should have 3 characters, remaining on next line
-        assert!(
-            output.len() >= 2,
-            "Should have at least 2 lines when folding"
-        );
-        assert!(output[0].contains("test"));
-        assert!(output[0].contains("5")); // character count
-        assert!(
-            !output[1].trim().is_empty(),
-            "Second line should contain remaining characters"
-        );
-    }
-
-    #[test]
-    fn test_format_pinyin_output_alignment() {
-        let test_data = vec![
-            ("ji".to_string(), vec!["机".to_string()]),
-            ("longpinyin".to_string(), vec!["长".to_string()]),
-        ];
-
-        let output = format_pinyin_output(&test_data, None);
-
-        // Test output format
-        for line in &output {
-            assert!(line.contains(":"), "Each line should contain ':'");
-            let parts: Vec<&str> = line.split(':').collect();
-            assert_eq!(parts.len(), 2, "Each line should have exactly one ':'");
-        }
-    }
-
-    #[test]
-    fn test_group_by_tone_found() {
-        let records = create_test_records();
-        let result = group_by_tone(&records, "ji", false);
-
-        assert!(result.is_some());
-        let tone_groups = result.unwrap();
-
-        // ji has 2 characters (tone 1: 机, tone 4: 计)
-        assert_eq!(tone_groups.len(), 2);
-
-        // Should be sorted by tone order
-        assert_eq!(tone_groups[0].0, 1); // tone 1
-        assert_eq!(tone_groups[0].2, vec!["机"]); // 机
-
-        assert_eq!(tone_groups[1].0, 4); // tone 4
-        assert_eq!(tone_groups[1].2, vec!["计"]); // 计
-    }
-
-    #[test]
-    fn test_group_by_tone_traditional() {
-        let records = create_test_records();
-        let result = group_by_tone(&records, "ji", true);
-
-        assert!(result.is_some());
-        let tone_groups = result.unwrap();
-
-        // Traditional characters should be used
-        assert_eq!(tone_groups[0].2, vec!["機"]); // 機 (traditional)
-        assert_eq!(tone_groups[1].2, vec!["計"]); // 計 (traditional)
-    }
-
-    #[test]
-    fn test_group_by_tone_not_found() {
-        let records = create_test_records();
-        let result = group_by_tone(&records, "nonexistent", false);
-
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_group_by_tone_pinyin_with_tone_marks() {
-        let records = create_test_records();
-        let result = group_by_tone(&records, "ji", false);
-
-        assert!(result.is_some());
-        let tone_groups = result.unwrap();
-
-        // pinyin should contain tone marks
-        assert_eq!(tone_groups[0].1, "jī"); // tone 1
-        assert_eq!(tone_groups[1].1, "jì"); // tone 4
-    }
-
-    #[test]
-    fn test_format_tone_output() {
-        let test_data = vec![
-            (1, "jī".to_string(), vec!["机".to_string()]),
-            (
-                4,
-                "jì".to_string(),
-                vec!["计".to_string(), "记".to_string()],
-            ),
-        ];
-
-        let output = format_tone_output(&test_data);
-
-        assert_eq!(output.len(), 2);
-        assert_eq!(output[0], "jī: 机");
-        assert_eq!(output[1], "jì: 计记");
-    }
-
-    #[test]
-    fn test_format_tone_output_empty() {
-        let test_data = vec![];
-        let output = format_tone_output(&test_data);
-
-        assert!(output.is_empty());
-    }
-
-    #[test]
-    fn test_tone_sorting() {
-        let mut records = create_test_records();
-        // Additional test data: tone 5 (neutral tone)
-        records.push(HanziRecord {
-            frequency: 4,
-            simplified: "吗".to_string(),
-            traditional: "嗎".to_string(),
-            pinyin: "ma".to_string(),
-            pinyin_without_tone: "ma".to_string(),
-            tone: 5, // neutral tone
-            onset: HanziOnset::M,
-            rime: HanziRime::A,
-        });
-
-        let result = group_by_tone(&records, "ma", false);
-        assert!(result.is_some());
-        let tone_groups = result.unwrap();
-
-        // Should contain tone 3 (马) and tone 5 (吗)
-        assert_eq!(tone_groups.len(), 2);
-        assert_eq!(tone_groups[0].0, 3); // tone 3 comes first
-        assert_eq!(tone_groups[1].0, 5); // tone 5 comes after
-    }
-
-    #[test]
-    fn test_hanzi_onset_as_str() {
-        // Test single character onsets
-        assert_eq!(HanziOnset::B.as_str(), "b");
-        assert_eq!(HanziOnset::P.as_str(), "p");
-        assert_eq!(HanziOnset::M.as_str(), "m");
-        assert_eq!(HanziOnset::F.as_str(), "f");
-        assert_eq!(HanziOnset::D.as_str(), "d");
-        assert_eq!(HanziOnset::T.as_str(), "t");
-        assert_eq!(HanziOnset::N.as_str(), "n");
-        assert_eq!(HanziOnset::Z.as_str(), "z");
-        assert_eq!(HanziOnset::C.as_str(), "c");
-        assert_eq!(HanziOnset::S.as_str(), "s");
-        assert_eq!(HanziOnset::L.as_str(), "l");
-        assert_eq!(HanziOnset::R.as_str(), "r");
-        assert_eq!(HanziOnset::J.as_str(), "j");
-        assert_eq!(HanziOnset::Q.as_str(), "q");
-        assert_eq!(HanziOnset::X.as_str(), "x");
-        assert_eq!(HanziOnset::G.as_str(), "g");
-        assert_eq!(HanziOnset::K.as_str(), "k");
-        assert_eq!(HanziOnset::H.as_str(), "h");
-        assert_eq!(HanziOnset::Y.as_str(), "y");
-        assert_eq!(HanziOnset::W.as_str(), "w");
-
-        // Test compound onsets (kebab-case format)
-        assert_eq!(HanziOnset::Zh.as_str(), "zh");
-        assert_eq!(HanziOnset::Ch.as_str(), "ch");
-        assert_eq!(HanziOnset::Sh.as_str(), "sh");
-
-        // Test None case
-        assert_eq!(HanziOnset::None.as_str(), "none");
-    }
 
     #[test]
     fn test_pinyin_v_to_u_replacement() {
@@ -1347,212 +73,5 @@ mod tests {
             result_with_u.is_some(),
             "Search with 'ü' should find characters"
         );
-    }
-
-    #[test]
-    fn test_onset_from_str() {
-        use std::str::FromStr;
-
-        // Test valid single-character onsets
-        assert_eq!(HanziOnset::from_str("b"), Ok(HanziOnset::B));
-        assert_eq!(HanziOnset::from_str("B"), Ok(HanziOnset::B));
-        assert_eq!(HanziOnset::from_str("p"), Ok(HanziOnset::P));
-        assert_eq!(HanziOnset::from_str("m"), Ok(HanziOnset::M));
-
-        // Test valid multi-character onsets
-        assert_eq!(HanziOnset::from_str("zh"), Ok(HanziOnset::Zh));
-        assert_eq!(HanziOnset::from_str("Zh"), Ok(HanziOnset::Zh));
-        assert_eq!(HanziOnset::from_str("ZH"), Ok(HanziOnset::Zh));
-        assert_eq!(HanziOnset::from_str("ch"), Ok(HanziOnset::Ch));
-        assert_eq!(HanziOnset::from_str("sh"), Ok(HanziOnset::Sh));
-
-        // Test special case
-        assert_eq!(HanziOnset::from_str("none"), Ok(HanziOnset::None));
-        assert_eq!(HanziOnset::from_str("None"), Ok(HanziOnset::None));
-        assert_eq!(HanziOnset::from_str("NONE"), Ok(HanziOnset::None));
-
-        // Test invalid inputs
-        assert!(HanziOnset::from_str("invalid").is_err());
-        assert!(HanziOnset::from_str("").is_err());
-        assert!(HanziOnset::from_str("zz").is_err());
-
-        // Test error message
-        let result = HanziOnset::from_str("invalid");
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "Invalid onset: 'invalid'");
-    }
-
-    #[test]
-    fn test_set_hanzi_onsets_refactored() {
-        // Test the refactored set_hanzi_onsets function with specific cases
-        let mut test_records = vec![
-            HanziRecord {
-                frequency: 1,
-                simplified: "中".to_string(),
-                traditional: "中".to_string(),
-                pinyin: "zhōng".to_string(),
-                pinyin_without_tone: "zhong".to_string(),
-                tone: 1,
-                onset: HanziOnset::None, // Initial value
-                rime: HanziRime::None,
-            },
-            HanziRecord {
-                frequency: 2,
-                simplified: "是".to_string(),
-                traditional: "是".to_string(),
-                pinyin: "shì".to_string(),
-                pinyin_without_tone: "shi".to_string(),
-                tone: 4,
-                onset: HanziOnset::None, // Initial value
-                rime: HanziRime::None,
-            },
-            HanziRecord {
-                frequency: 3,
-                simplified: "马".to_string(),
-                traditional: "马".to_string(),
-                pinyin: "mǎ".to_string(),
-                pinyin_without_tone: "ma".to_string(),
-                tone: 3,
-                onset: HanziOnset::None, // Initial value
-                rime: HanziRime::None,
-            },
-            HanziRecord {
-                frequency: 4,
-                simplified: "安".to_string(),
-                traditional: "安".to_string(),
-                pinyin: "ān".to_string(),
-                pinyin_without_tone: "an".to_string(),
-                tone: 1,
-                onset: HanziOnset::None, // Initial value
-                rime: HanziRime::None,
-            },
-        ];
-
-        // Apply the refactored set_hanzi_onsets function
-        set_hanzi_onsets(&mut test_records);
-
-        // Verify the results
-        assert_eq!(test_records[0].onset, HanziOnset::Zh); // "zhong" -> Zh
-        assert_eq!(test_records[1].onset, HanziOnset::Sh); // "shi" -> Sh
-        assert_eq!(test_records[2].onset, HanziOnset::M); // "ma" -> M
-        assert_eq!(test_records[3].onset, HanziOnset::None); // "an" -> None (vowel-initial)
-    }
-
-    #[test]
-    fn test_hanzi_rime_as_str() {
-        // Test simple vowel rimes
-        assert_eq!(HanziRime::E.as_str(), "e");
-        assert_eq!(HanziRime::A.as_str(), "a");
-        assert_eq!(HanziRime::I.as_str(), "i");
-        assert_eq!(HanziRime::U.as_str(), "u");
-
-        // Test compound vowel rimes
-        assert_eq!(HanziRime::Ei.as_str(), "ei");
-        assert_eq!(HanziRime::Ai.as_str(), "ai");
-        assert_eq!(HanziRime::Ou.as_str(), "ou");
-        assert_eq!(HanziRime::Ao.as_str(), "ao");
-
-        // Test nasal ending rimes
-        assert_eq!(HanziRime::En.as_str(), "en");
-        assert_eq!(HanziRime::An.as_str(), "an");
-        assert_eq!(HanziRime::Eng.as_str(), "eng");
-        assert_eq!(HanziRime::Ang.as_str(), "ang");
-        assert_eq!(HanziRime::Ong.as_str(), "ong");
-
-        // Test complex compound rimes
-        assert_eq!(HanziRime::Ie.as_str(), "ie");
-        assert_eq!(HanziRime::Ia.as_str(), "ia");
-        assert_eq!(HanziRime::Iu.as_str(), "iu");
-        assert_eq!(HanziRime::Iao.as_str(), "iao");
-        assert_eq!(HanziRime::In.as_str(), "in");
-        assert_eq!(HanziRime::Ian.as_str(), "ian");
-        assert_eq!(HanziRime::Iong.as_str(), "iong");
-        assert_eq!(HanziRime::Ing.as_str(), "ing");
-        assert_eq!(HanziRime::Iang.as_str(), "iang");
-
-        // Test u-compound rimes
-        assert_eq!(HanziRime::Uo.as_str(), "uo");
-        assert_eq!(HanziRime::Ua.as_str(), "ua");
-        assert_eq!(HanziRime::Ui.as_str(), "ui");
-        assert_eq!(HanziRime::Uai.as_str(), "uai");
-        assert_eq!(HanziRime::Un.as_str(), "un");
-        assert_eq!(HanziRime::Uan.as_str(), "uan");
-        assert_eq!(HanziRime::Uang.as_str(), "uang");
-
-        // Test special ü rimes
-        assert_eq!(HanziRime::V.as_str(), "ü");
-        assert_eq!(HanziRime::Ve.as_str(), "üe");
-
-        // Test None case
-        assert_eq!(HanziRime::None.as_str(), "none");
-    }
-
-    #[test]
-    fn test_rime_from_str() {
-        use std::str::FromStr;
-
-        // Test valid simple vowel rimes
-        assert_eq!(HanziRime::from_str("e"), Ok(HanziRime::E));
-        assert_eq!(HanziRime::from_str("a"), Ok(HanziRime::A));
-        assert_eq!(HanziRime::from_str("i"), Ok(HanziRime::I));
-        assert_eq!(HanziRime::from_str("u"), Ok(HanziRime::U));
-
-        // Test valid compound vowel rimes
-        assert_eq!(HanziRime::from_str("ei"), Ok(HanziRime::Ei));
-        assert_eq!(HanziRime::from_str("ai"), Ok(HanziRime::Ai));
-        assert_eq!(HanziRime::from_str("ou"), Ok(HanziRime::Ou));
-        assert_eq!(HanziRime::from_str("ao"), Ok(HanziRime::Ao));
-
-        // Test valid nasal ending rimes
-        assert_eq!(HanziRime::from_str("en"), Ok(HanziRime::En));
-        assert_eq!(HanziRime::from_str("an"), Ok(HanziRime::An));
-        assert_eq!(HanziRime::from_str("eng"), Ok(HanziRime::Eng));
-        assert_eq!(HanziRime::from_str("ang"), Ok(HanziRime::Ang));
-        assert_eq!(HanziRime::from_str("ong"), Ok(HanziRime::Ong));
-
-        // Test valid complex compound rimes
-        assert_eq!(HanziRime::from_str("ie"), Ok(HanziRime::Ie));
-        assert_eq!(HanziRime::from_str("ia"), Ok(HanziRime::Ia));
-        assert_eq!(HanziRime::from_str("iu"), Ok(HanziRime::Iu));
-        assert_eq!(HanziRime::from_str("iao"), Ok(HanziRime::Iao));
-        assert_eq!(HanziRime::from_str("in"), Ok(HanziRime::In));
-        assert_eq!(HanziRime::from_str("ian"), Ok(HanziRime::Ian));
-        assert_eq!(HanziRime::from_str("iong"), Ok(HanziRime::Iong));
-        assert_eq!(HanziRime::from_str("ing"), Ok(HanziRime::Ing));
-        assert_eq!(HanziRime::from_str("iang"), Ok(HanziRime::Iang));
-
-        // Test valid u-compound rimes
-        assert_eq!(HanziRime::from_str("uo"), Ok(HanziRime::Uo));
-        assert_eq!(HanziRime::from_str("ua"), Ok(HanziRime::Ua));
-        assert_eq!(HanziRime::from_str("ui"), Ok(HanziRime::Ui));
-        assert_eq!(HanziRime::from_str("uai"), Ok(HanziRime::Uai));
-        assert_eq!(HanziRime::from_str("un"), Ok(HanziRime::Un));
-        assert_eq!(HanziRime::from_str("uan"), Ok(HanziRime::Uan));
-        assert_eq!(HanziRime::from_str("uang"), Ok(HanziRime::Uang));
-
-        // Test special ü rimes - both ü and v representations
-        assert_eq!(HanziRime::from_str("ü"), Ok(HanziRime::V));
-        assert_eq!(HanziRime::from_str("v"), Ok(HanziRime::V));
-        assert_eq!(HanziRime::from_str("üe"), Ok(HanziRime::Ve));
-        assert_eq!(HanziRime::from_str("ve"), Ok(HanziRime::Ve));
-
-        // Test case insensitivity
-        assert_eq!(HanziRime::from_str("ANG"), Ok(HanziRime::Ang));
-        assert_eq!(HanziRime::from_str("Iang"), Ok(HanziRime::Iang));
-
-        // Test special case
-        assert_eq!(HanziRime::from_str("none"), Ok(HanziRime::None));
-        assert_eq!(HanziRime::from_str("None"), Ok(HanziRime::None));
-        assert_eq!(HanziRime::from_str("NONE"), Ok(HanziRime::None));
-
-        // Test invalid inputs
-        assert!(HanziRime::from_str("invalid").is_err());
-        assert!(HanziRime::from_str("").is_err());
-        assert!(HanziRime::from_str("xyz").is_err());
-
-        // Test error message
-        let result = HanziRime::from_str("invalid");
-        assert!(result.is_err());
-        assert_eq!(result.unwrap_err(), "Invalid rime: 'invalid'");
     }
 }
